refactor(#233): per-team TeammateCollection + src/service restructure
Align the code with the accepted design after an implementation deviation
(a single shared TeammateCollection discriminated by a teamId argument).

Per-team model:
- Each TeammateCollection bakes a fixed teamScope (null = dispatcher-owned,
  team_id = a team's leader + members); public methods drop the teamId arg and
  feed the baked scope to the unchanged stores/read-model.
- TeamService OWNS its per-team TeammateCollection. TeamCollection.get is a
  get-or-rebuild factory (cached by team_id, rebuilt from TeamRecord on miss),
  uniform with Dispatchers.get and TeammateCollection.entityFor; dissolve evicts
  the cache, and all TeamRecord mutations route through the cached service.
- Names stay dispatcher-global (allocateName -> listAllNames), so the
  CompletionRouter key stays producerName:turnId with no team_id; delivery
  topology stays solely in DispatcherService.initiatorFor.
- shutdown() sweeps only cached collections (dispatcher + each materialized
  team's members/leader) before the dispatcher agent.

Restructure:
- Delete the top-level src/dispatcher-service/; move the module under
  src/service/ with one service class per file/dir (helpers colocated, shared
  helpers at the service root). Split the two-class team/service.ts into
  team-collection/ and team-service/. Add a service/index.ts barrel and fix all
  importers (server, admin, mcp, cli, tests).

State layout is an incompatible 0.x break (fail-loud, no migration); ships a
Rush minor change file with BREAKING/Rebuild notes. Gates: tsc 0, lint 0
(no-sync-IO), rush rebuild 0 (8/8), 486 tests pass.
