+++ .agents/reference/state-and-paths.md
@@ -50,11 +50,11 @@ would live under `~/.dreamux`.
 
 Key source:
 
-- `/packages/dreamux/src/dispatcher-service/dispatcher-workspace.ts`
-- `/packages/dreamux/src/dispatcher-service/teammate/workspaces.ts`
-- `/packages/dreamux/src/dispatcher-service/teammate/worktree-manager.ts`
-- `/packages/dreamux/src/dispatcher-service/teammate/worktree-paths.ts`
-- `/packages/dreamux/src/dispatcher-service/team/service.ts`
+- `/packages/dreamux/src/service/dispatcher-service/workspace.ts`
+- `/packages/dreamux/src/service/worktree/workspaces.ts`
+- `/packages/dreamux/src/service/worktree/manager.ts`
+- `/packages/dreamux/src/service/worktree/paths.ts`
+- `/packages/dreamux/src/service/team-collection/index.ts`
 
 ## Server-Owned State
 
@@ -86,9 +86,9 @@ Key source:
 
 - `/packages/dreamux/src/platform/paths.ts`
 - `/packages/dreamux/src/state/dispatcher-store.ts`
-- `/packages/dreamux/src/dispatcher-service/teammate/identity-store.ts`
-- `/packages/dreamux/src/dispatcher-service/teammate/turns-store.ts`
-- `/packages/dreamux/src/dispatcher-service/team/store.ts`
+- `/packages/dreamux/src/service/teammate-collection/identity-store.ts`
+- `/packages/dreamux/src/service/teammate-collection/turns-store.ts`
+- `/packages/dreamux/src/service/team-collection/store.ts`
 - `/packages/channel/feishu-channel/src/chat-bots-store.ts`
 
 ## Run Files
