+++ packages/dreamux/CLAUDE.md
@@ -32,9 +32,9 @@ Two settled shape rules govern where code lives:
 |---|---|---|
 | `server.ts` | process entry + wiring only | builds registry/catalog/store/services, opens the admin socket, starts dispatchers; owns no teammate/channel/runtime orchestration |
 | `agent-runtime/` | the AgentRuntime seam, all neutral: `catalog.ts` (registry-backed `AgentRuntimeProviderCatalog` + builtin registration), `external-provider.ts` loader, `load-config.ts` (composes builtin agentRuntime + channel registration), `host-context.ts` / `host-paths.ts` (host adapters that bridge core's store/logger/path layout onto the neutral `@excitedjs/dreamux-types` create context), `bundled-skill-sources.ts`, `index.ts` barrel. Contract types live in `@excitedjs/dreamux-types` — core imports them directly (no in-core `types.ts` / `turn.ts`) | one neutral abstraction for every agent role; core never names a concrete runtime |
-| `dispatcher-service/` | the Dispatcher Service entity — see [`dispatcher-service/CLAUDE.md`](src/dispatcher-service/CLAUDE.md) | holds the dispatcher agent + orchestrates teammates |
-| `dispatcher-service/dispatcher/` | the dispatcher agent as a contained `TeammateService` (`agent.ts` factory) + the live `ChannelSessions` (`channel-sessions.ts`) + the TeamLeader channel-egress gate (`channel-tool-auth.ts`) + role MCP descriptor builders (`mcp-descriptors.ts`) + dispatcher base prompt + runnable-channel guard. Agent runtime lifecycle (start/resume/stop) lives in the shared `TeammateService`; `DispatcherService` (`dispatcher-instance.ts`) keeps channel sessions, restart-notice injection, MCP assembly, and completion routing | the dispatcher *has an* agent (Phase 5, #233); there is no separate `DispatcherRuntimeService` |
-| `dispatcher-service/teammate/` | `TeammateCollection` + `TeammateService` entity + `CompletionRouter` + identity-store + runtime-state + types + teammate MCP descriptor | agent-centric teammates (no `task`): spawn/send/close + forward-only history (send reopens a closed teammate; no separate `resume` verb, #155) |
+| `service/` | the Dispatcher Service module (issue #233 restructure): one service class per file/dir, with `index.ts` the package-internal barrel — see [`service/CLAUDE.md`](src/service/CLAUDE.md) | holds the dispatcher agent + orchestrates teammates |
+| `service/dispatcher-service/` | the per-dispatcher aggregate (`index.ts` = `DispatcherService`) + its agent-side parts: the dispatcher agent as a contained `TeammateService` (`agent.ts` factory), the live `ChannelSessions` (`channel-sessions.ts`), the TeamLeader channel-egress gate (`channel-tool-auth.ts`), role MCP descriptor builders (`mcp-descriptors.ts`), dispatcher base prompt, runnable-channel guard, workspace policy (`workspace.ts`), and `errors.ts`. Agent runtime lifecycle (start/resume/stop) lives in the shared `TeammateService`; `DispatcherService` keeps channel sessions, restart-notice injection, MCP assembly, and completion routing. `service/dispatchers/` holds the process-level `Dispatchers` collection | the dispatcher *has an* agent (Phase 5, #233); there is no separate `DispatcherRuntimeService` |
+| `service/teammate-collection/` (+ `service/teammate-service/`, `service/completion-router/`) | `TeammateCollection` (collection: stores, factory paths, router registration) + `TeammateService` entity + `CompletionRouter` + identity/turns stores + read-model + runtime-state + types + teammate MCP descriptor. `service/team-collection/` + `service/team-service/` hold `TeamCollection` / `TeamService`; `service/worktree/`, `service/channel-binding/`, `service/legacy-state.ts` are shared helpers at the module root | agent-centric teammates (no `task`): spawn/send/close + forward-only history (send reopens a closed teammate; no separate `resume` verb, #155) |
 | `channel/` | the bidirectional conversational ChannelProvider seam, all neutral: `catalog.ts` (`ChannelProviderCatalog`) and `external-channel-provider.ts` loader | the channel engine (session/bot/gate/message/tool-parsing/identity) lives in the provider package and never imports core; core is a blind channel-MCP conduit — the generic `channel.invoke_tool` admin method routed to the neutral `ChannelSession` / `ChannelProvider` seam — plus neutral routing/binding/auth |
 | `@excitedjs/dreamux-types` `SubscribeChannelProvider` | public TS contract reservation for future one-way subscription channels (GitHub/Jira issue or PR feeds) | type-only this phase; core has no runnable loader/catalog yet. Subscription channels do not have `chat_id`, Team binding, reply/react, or conversational target ownership |
 | `registry/` | provider registry/loader + provider-ref grammar | resolves `builtin:` / `npm:` refs; runnable catalogs currently cover `channel` and `agentRuntime`; `subscribeChannel` is a type-level reservation until a subscription loader/lifecycle exists |
