+++ packages/dreamux/src/service/index.ts
@@ -1,0 +1,4 @@
+export { Dispatchers, type DispatchersOptions } from './dispatchers/index.js';
+export { DispatcherService } from './dispatcher-service/index.js';
+export { TeamService } from './team-service/index.js';
+export { ChannelToolAuthorizationError } from './dispatcher-service/errors.js';
