+++ packages/dreamux/src/service/team-collection/mcp-config.ts
