+++ packages/dreamux/src/service/teammate-collection/mcp-config.ts
